upgrade pinecone to v1
