+++ packages/components/nodes/vectorstores/Pinecone/Pinecone_Upsert.ts
@@ -1,10 +1,12 @@
 import { ICommonObject, INode, INodeData, INodeOutputsValue, INodeParams } from '../../../src/Interface'
-import { PineconeClient } from '@pinecone-database/pinecone'
+import { Pinecone } from '@pinecone-database/pinecone'
 import { PineconeLibArgs, PineconeStore } from 'langchain/vectorstores/pinecone'
 import { Embeddings } from 'langchain/embeddings/base'
+import { Document } from 'langchain/document'
 import { getBaseClasses, getCredentialData, getCredentialParam } from '../../../src/utils'
+import { flatten } from 'lodash'
 
-class Pinecone_Existing_VectorStores implements INode {
+class PineconeUpsert_VectorStores implements INode {
     label: string
     name: string
     version: number
@@ -18,13 +20,13 @@ class Pinecone_Existing_VectorStores implements INode {
     outputs: INodeOutputsValue[]
 
     constructor() {
-        this.label = 'Pinecone Load Existing Index'
-        this.name = 'pineconeExistingIndex'
+        this.label = 'Pinecone Upsert Document'
+        this.name = 'pineconeUpsert'
         this.version = 1.0
         this.type = 'Pinecone'
         this.icon = 'pinecone.png'
         this.category = 'Vector Stores'
-        this.description = 'Load existing index from Pinecone (i.e: Document has been upserted)'
+        this.description = 'Upsert documents to Pinecone'
         this.baseClasses = [this.type, 'VectorStoreRetriever', 'BaseRetriever']
         this.credential = {
             label: 'Connect Credential',
@@ -33,6 +35,12 @@ class Pinecone_Existing_VectorStores implements INode {
             credentialNames: ['pineconeApi']
         }
         this.inputs = [
+            {
+                label: 'Document',
+                name: 'document',
+                type: 'Document',
+                list: true
+            },
             {
                 label: 'Embeddings',
                 name: 'embeddings',
@@ -50,13 +58,6 @@ class Pinecone_Existing_VectorStores implements INode {
                 placeholder: 'my-first-namespace',
                 additionalParams: true,
                 optional: true
-            },
-            {
-                label: 'Pinecone Metadata Filter',
-                name: 'pineconeMetadataFilter',
-                type: 'json',
-                optional: true,
-                additionalParams: true
             },
             {
                 label: 'Top K',
@@ -85,7 +86,7 @@ class Pinecone_Existing_VectorStores implements INode {
     async init(nodeData: INodeData, _: string, options: ICommonObject): Promise<any> {
         const index = nodeData.inputs?.pineconeIndex as string
         const pineconeNamespace = nodeData.inputs?.pineconeNamespace as string
-        const pineconeMetadataFilter = nodeData.inputs?.pineconeMetadataFilter
+        const docs = nodeData.inputs?.document as Document[]
         const embeddings = nodeData.inputs?.embeddings as Embeddings
         const output = nodeData.outputs?.output as string
         const topK = nodeData.inputs?.topK as string
@@ -95,25 +96,26 @@ class Pinecone_Existing_VectorStores implements INode {
         const pineconeApiKey = getCredentialParam('pineconeApiKey', credentialData, nodeData)
         const pineconeEnv = getCredentialParam('pineconeEnv', credentialData, nodeData)
 
-        const client = new PineconeClient()
-        await client.init({
+        const client = new Pinecone({
             apiKey: pineconeApiKey,
             environment: pineconeEnv
         })
 
         const pineconeIndex = client.Index(index)
 
+        const flattenDocs = docs && docs.length ? flatten(docs) : []
+        const finalDocs = []
+        for (let i = 0; i < flattenDocs.length; i += 1) {
+            finalDocs.push(new Document(flattenDocs[i]))
+        }
+
         const obj: PineconeLibArgs = {
             pineconeIndex
         }
 
         if (pineconeNamespace) obj.namespace = pineconeNamespace
-        if (pineconeMetadataFilter) {
-            const metadatafilter = typeof pineconeMetadataFilter === 'object' ? pineconeMetadataFilter : JSON.parse(pineconeMetadataFilter)
-            obj.filter = metadatafilter
-        }
 
-        const vectorStore = await PineconeStore.fromExistingIndex(embeddings, obj)
+        const vectorStore = await PineconeStore.fromDocuments(finalDocs, embeddings, obj)
 
         if (output === 'retriever') {
             const retriever = vectorStore.asRetriever(k)
@@ -126,4 +128,4 @@ class Pinecone_Existing_VectorStores implements INode {
     }
 }
 
-module.exports = { nodeClass: Pinecone_Existing_VectorStores }
+module.exports = { nodeClass: PineconeUpsert_VectorStores }
