+++ packages/components/nodes/vectorstores/Pinecone/pinecone.png
